// src/api/models/HandlerResponse_AudioResponse_.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { AudioResponse } from './AudioResponse';

export type HandlerResponse_AudioResponse_ = {
  entities: Array<AudioResponse>;
  id: string;
  name: string;
};