// src/api/models/HandlerResponse_LightsGroupResponse_.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { LightsGroupResponse } from './LightsGroupResponse';

export type HandlerResponse_LightsGroupResponse_ = {
  entities: Array<LightsGroupResponse>;
  id: string;
  name: string;
};