// src/api/models/HandlerResponse_ScreenResponse_.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { ScreenResponse } from './ScreenResponse';

export type HandlerResponse_ScreenResponse_ = {
  entities: Array<ScreenResponse>;
  id: string;
  name: string;
};